src/pages/segmenter.py
import streamlit as st
from segment_anything import sam_model_registry, SamPredictor
import numpy as np
from PIL import Image

st.title("Image Segmenter")

uploaded_file = st.file_uploader("Choose an image...", type=["jpg", "jpeg", "png"])

if uploaded_file is not None:
    image = Image.open(uploaded_file)
    st.image(image, caption="Original Image")
    
    if st.button("Segment Image"):
        # SAM 모델 구현
        # 실제 구현시에는 모델 다운로드 및 초기화 로직이 필요합니다
        st.write("Segmentation will be implemented here") 
